+++ packages/dsuite-core/src/parties/documents.js
@@ -2,10 +2,10 @@
 // Copyright 2019 Wireline, Inc.
 //
 
+const PartyManager = require('../parties/party_manager');
 const { encodeFeedKey, decodeFeedKey } = require('../protocol/feeds');
 
-// TODO(burdon): Remove dsuite dependency.
-module.exports = function setRules(dsuite) {
+module.exports = ({ core, mega, partyManager }) => {
   return {
     name: 'dsuite:documents',
 
@@ -15,7 +15,7 @@ module.exports = function setRules(dsuite) {
     },
 
     async getParticipantKeys(partyKey) {
-      const participants = await dsuite.api['participants'].getParticipants({ partyKey });
+      const participants = await core.api['participants'].getParticipants({ partyKey });
 
       return participants.reduce((prev, participant) => (
         [
@@ -27,11 +27,11 @@ module.exports = function setRules(dsuite) {
     },
 
     async handshake({ peer }) {
-      dsuite.emit(`rule:${this.name}:handshake`, { rule: this, peer });
+      partyManager.emit('rule-handshake', { rule: this, peer });
 
       const partyKey = peer.party.key.toString('hex');
-      const controlFeed = dsuite.mega.feed('control');
-      const feed = await dsuite.getLocalPartyFeed(partyKey);
+      const controlFeed = mega.feed('control');
+      const feed = await partyManager.getLocalPartyFeed(partyKey);
 
       const participantKeys = await this.getParticipantKeys(partyKey);
 
@@ -46,7 +46,7 @@ module.exports = function setRules(dsuite) {
 
       keys.forEach((feedKey) => {
         const { key } = decodeFeedKey(feedKey);
-        const feed = dsuite.mega.feed(key);
+        const feed = mega.feed(key);
         peer.replicate(feed);
       });
 
@@ -64,19 +64,19 @@ module.exports = function setRules(dsuite) {
 
         keys.forEach((feedKey) => {
           const { key } = decodeFeedKey(feedKey);
-          const feed = dsuite.mega.feed(key);
+          const feed = mega.feed(key);
           peer.replicate(feed);
         });
       };
 
-      dsuite.core.api['participants'].events.on('participant', onParticipant);
+      core.api['participants'].events.on('participant', onParticipant);
       peer.on('destroy', () => {
-        dsuite.core.api['participants'].events.removeListener('participant', onParticipant);
+        core.api['participants'].events.removeListener('participant', onParticipant);
       });
     },
 
     async onEphemeralMessage({ message, peer }) {
-      dsuite.emit(`rule:${this.name}:message`, { rule: this, message, peer });
+      partyManager.emit('rule-ephemeral-message', { rule: this, message, peer });
     },
 
     async onIntroduceFeeds({ message, peer }) {
@@ -88,12 +88,12 @@ module.exports = function setRules(dsuite) {
 
         let name;
         if (type === 'party-feed') {
-          name = dsuite.getPartyName(partyKey, key);
+          name = PartyManager.getPartyName(partyKey, key);
         } else {
           name = `control-feed/${key.toString('hex')}`;
         }
 
-        return dsuite.mega.addFeed({ name, key, load: false });
+        return mega.addFeed({ name, key, load: false });
       }));
 
       // The keys I allow to replicate.