+++ packages/dsuite-core/src/views/contacts.js
@@ -7,11 +7,11 @@ const view = require('kappa-view-level');
 const sub = require('subleveldown');
 
 const { streamToList } = require('../utils/stream');
+const { uuid } = require('../utils/uuid');
 const { append } = require('../protocol/messages');
 
 module.exports = function ContactsView(dsuite) {
-  const { uuid, mega, db } = dsuite;
-
+  const { mega, db } = dsuite;
   const events = new EventEmitter();
   events.setMaxListeners(Infinity);
 
@@ -65,7 +65,7 @@ module.exports = function ContactsView(dsuite) {
 
       // TODO(burdon): Comment.
       key() {
-        return dsuite.mega.key.toString('hex');
+        return mega.key.toString('hex');
       },
 
       // TODO(burdon): Query?